remove SwiftTypeOld
